add new metrics

+++ rosen/modules/render_service_base/src/platform/ohos/rs_marshalling_helper.cpp
@@ -3122,7 +3122,7 @@ bool RSMarshallingHelper::MarshallingTransactionVer(Parcel& parcel)
     for (auto supportedFlag : supportedParcelVerFlags) {
         flags[supportedFlag / bitsPerUint64] |= static_cast<uint64_t>(1) << (supportedFlag % bitsPerUint64);
     }
-    for (int i = 0; i < NUM_ITEMS_IN_VERSION; i++) {
+    for (size_t i = 0; i < NUM_ITEMS_IN_VERSION; i++) {
         parcel.WriteUint64(flags[i]);
     }
     return true;
@@ -3134,7 +3134,7 @@ bool RSMarshallingHelper::UnmarshallingTransactionVer(Parcel& parcel)
     size_t offset = parcel.GetReadPosition();
     headerCode = parcel.ReadInt64();
     if (headerCode == -1) {
-        for (int i = 0; i < NUM_ITEMS_IN_VERSION; i++) {
+        for (size_t i = 0; i < NUM_ITEMS_IN_VERSION; i++) {
             parcel.ReadUint64();
         }
     } else {
